+++ foodxpress-for-woocommerce.php
@@ -6,7 +6,8 @@
  * Version:           1.2.1
  * Author:            MD MILLAT HOSEN
  * Author URI:        https://github.com/codermillat
- * License:           Proprietary
+ * License:           GPL-3.0-or-later
+ * License URI:       https://www.gnu.org/licenses/gpl-3.0.html
  * Text Domain:       foodxpress
  * Domain Path:       /languages
  * Requires at least: 6.0